+++ examples/data-sources/fmc_endpoint_device_types/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_endpoint_device_types" "example" {
+  items = {
+    "my_endpoint_device_types" = {
+    }
+  }
+}